Lift team grid generation code out to its own class. Minimal changes to that code; just those required to class-ify it.
