+++ Forms/FormFixture.cs
@@ -56,15 +56,11 @@ namespace Torn.UI
 				printReport1.FileName = System.IO.Path.Combine(ExportFolder, holder.Key, "fixture." + holder.ReportTemplates.OutputFormat.ToExtension());
 		}
 
-		private List<List<int>> previousGrid;
-		private double previousBestScore;
-		private double previousGamesPerTeam;
-		private bool previousHasRef;
-		private List<List<int>> previousExistingPlays;
 		private List<CheckBox> teamSelectors = new List<CheckBox>();
 		private List<CheckBox> fixtureTeamSelectors = new List<CheckBox>();
 		private List<LeagueTeam> selectedTeams = new List<LeagueTeam>();
 		private List<LeagueTeam> fixtureSelectedTeams = new List<LeagueTeam>();
+		private TeamGrid teamGrid;
 
 		Colour leftButton, middleButton, rightButton, xButton1, xButton2;
 		Point point;  // This is the point in the grid last clicked on. It's counted in grid squares, not in pixels: 9,9 is ninth column, ninth row.
@@ -99,468 +95,6 @@ namespace Torn.UI
 			displayReportGrid.Report = null;
 		}
 
-		void LogGrid<T>(List<List<T>> grid)
-		{
-			string str = "**********\n";
-			foreach (List<T> list in grid)
-			{
-				foreach (T item in list)
-				{
-					str += item + ",\t";
-				}
-				str += "\n";
-			}
-			str += "**********";
-			Console.WriteLine(str);
-		}
-
-		List<T> FlattenGrid<T>(List<List<T>> grid)
-		{
-			List<T> flat = new List<T>();
-			foreach (List<T> list in grid)
-			{
-				flat.AddRange(list);
-			}
-			return flat;
-		}
-
-		List<List<T>> TransposeGrid<T>(List<List<T>> grid)
-		{
-			return grid.SelectMany(inner => inner.Select((item, index) => new { item, index }))
-				.GroupBy(i => i.index, i => i.item)
-				.Select(g => g.ToList())
-				.ToList();
-		}
-
-		List<List<int>> GetGrid(FixtureTeams teams, double teamsPerGame, double gamesPerTeam, bool hasRef, List<List<int>> existingPlays, int maxMillis)
-		{
-			List<List<int>> bestGrid = SetupGrid(teams, teamsPerGame, gamesPerTeam);
-
-			double bestScore = CalcScore(bestGrid, gamesPerTeam, hasRef, existingPlays);
-
-			return ContinueMixing(bestGrid, gamesPerTeam, hasRef, existingPlays, maxMillis, bestScore);
-
-		}
-
-		List<List<int>> ContinueMixing(List<List<int>> grid, double gamesPerTeam, bool hasRef, List<List<int>> existingPlays, int maxMillis, double startingScore)
-		{
-			List<List<int>> bestGrid = grid;
-			double bestScore = startingScore;
-			bool badFixture = bestScore > 10000;
-
-			Console.WriteLine("initScore: {0}", bestScore);
-			LogGrid(bestGrid);
-
-			Stopwatch sw = new Stopwatch();
-			sw.Start();
-			scoreLabel.Visible = false;
-			scoreLabel.Text = "";
-			int count = 0;
-
-			for (int i = 0; i < 20000; i++)
-			{
-				count++;
-				List<List<int>> mixedGrid = MixGrid(bestGrid);
-				double score = CalcScore(mixedGrid, gamesPerTeam, hasRef, existingPlays);
-				if (score <= bestScore)
-				{
-					bestScore = score;
-					bestGrid = mixedGrid;
-					if (bestScore > 10000)
-					{
-						badFixture = true;
-					}
-					else
-					{
-						badFixture = false;
-					}
-				} 
-			}
-
-			while (badFixture)
-			{
-				count++;
-				if (sw.ElapsedMilliseconds > maxMillis)
-					break;
-				List<List<int>> mixedGrid = MixGrid(bestGrid);
-				double score = CalcScore(mixedGrid, gamesPerTeam, hasRef, existingPlays);
-				if (score <= bestScore)
-				{
-					bestScore = score;
-					bestGrid = mixedGrid;
-					if (bestScore > 10000)
-					{
-						badFixture = true;
-					}
-					else
-					{
-						badFixture = false;
-					}
-				}
-			}
-			CalcScore(bestGrid, gamesPerTeam, hasRef, existingPlays, true);
-			Console.WriteLine("Time Elapsed (ms): {0}", sw.ElapsedMilliseconds);
-			Console.WriteLine("Iterations: {0}", count);
-
-			Console.WriteLine("bestScore: {0}", bestScore);
-			LogGrid(bestGrid);
-			Console.WriteLine("Existing Plays");
-			LogGrid(NormalisePlays(existingPlays));
-			Console.WriteLine("Plays");
-			LogGrid(NormalisePlays(CalcPlays(bestGrid, hasRef, existingPlays)));
-			scoreLabel.Text = bestScore >= 10000 ? "Extend max time and regenerate: " + Math.Round(bestScore).ToString() : "Score: " + Math.Round(bestScore).ToString() + " (Lower is better)";
-			scoreLabel.BackColor = bestScore >= 10000 ? Color.FromName("red") : Color.Transparent;
-			scoreLabel.Visible = true;
-
-			previousGrid = bestGrid;
-			previousBestScore = bestScore;
-			previousGamesPerTeam = gamesPerTeam;
-			previousHasRef = hasRef;
-			previousExistingPlays = existingPlays;
-			continueGenerating.Enabled = true;
-
-
-			return bestGrid;
-		}
-
-		List<List<int>> MixGrid(List<List<int>> grid)
-		{
-			List<List<int>> newGrid = grid.ConvertAll(row => row.ConvertAll(cell => cell));
-			double teamsPerGame = grid[0].Count;
-			double totalTeams = FlattenGrid(grid).Uniq().Count;
-			Random rnd = new Random();
-
-			int game1 = rnd.Next(grid.Count);
-			int game2 = rnd.Next(grid.Count);
-
-			int player1 = rnd.Next((int)teamsPerGame);
-			int player2 = rnd.Next((int)teamsPerGame);
-
-			newGrid[game1][player1] = grid[game2][player2];
-			newGrid[game2][player2] = grid[game1][player1];
-
-			return newGrid;
-		}
-
-		List<int> SumRows(List<List<int>> grid )
-		{
-			List<int> totals = new List<int>();
-			foreach(List<int> row in grid)
-			{
-				int total = 0;
-				foreach(int value in row)
-				{
-					total += value;
-				}
-				totals.Add(total);
-			}
-			return totals;
-		}
-
-		int AverageRow(List<int> row)
-		{
-			int total = 0;
-			foreach (int value in row)
-			{
-				total += value;
-			}
-
-			return total / row.Count;
-				
-		}
-
-		List<List<int>> NormalisePlays(List<List<int>> grid)
-		{
-            List<int> playsTotals = SumRows(grid);
-            double mostPlays = 0;
-			List<List<int>> updatedGrid = new List<List<int>>();
-
-            foreach (int totalPlays in playsTotals)
-            {
-                if (totalPlays > mostPlays)
-                {
-                    mostPlays = totalPlays;
-                }
-            }
-
-
-			for (int i = 0; i < grid.Count; i++)
-			{
-				List<int> row = grid[i];
-				List<int> newRow = new List<int>();
-				double teamPlays = playsTotals[i];
-				double multiplier = teamPlays == 0 || mostPlays == 0 ? 1 : mostPlays / teamPlays;
-
-				for(int j =  0; j < row.Count; j++)
-				{
-                    double teamPlays2 = playsTotals[j];
-                    double multiplier2 = teamPlays2 == 0 || mostPlays == 0 ? 1 : mostPlays / teamPlays2;
-					double mult = Math.Max(multiplier, multiplier2);
-					double result = mult * row[j];
-                    newRow.Add((int)Math.Round(result));
-				}
-				updatedGrid.Add(newRow);
-			}
-			return updatedGrid;
-        }
-
-		double GetAveragePlays(List<List<int>> grid)
-		{
-			double total = 0;
-            double count = 0;
-			for(int i = 0; i < grid.Count; i++)
-			{
-				for(int j = 0; j < grid[i].Count; j++)
-				{
-					if(i != j)
-					{
-						total += grid[i][j];
-						count++;
-					}
-				}
-			}
-			return total / count;
-		}
-
-		double CalcScore(List<List<int>> grid, double gamesPerTeam, bool hasRef, List<List<int>> existingPlays, bool log = false)
-		{
-			List<List<int>> normalisedExistingPlays = NormalisePlays(existingPlays);
-			int totalTeams = FlattenGrid(grid).Uniq().Count;
-			int BACK_TO_BACK_PENALTY = (int)backToBackPenalty.Value;
-			int teamsPerGame = grid[0].Count;
-			double previousAveragePlays = GetAveragePlays(normalisedExistingPlays);
-			List<List<int>> plays = NormalisePlays(CalcPlays(grid, hasRef, existingPlays));
-			double averagePlays = GetAveragePlays(plays);
-
-			double score = 0;
-
-			for (int player1 = 0; player1 < plays.Count; player1++)
-			{
-				score += plays[player1][player1] * 100000; // penalty for playing themselves
-				for(int player2 = player1 + 1; player2 < plays[player1].Count; player2++)
-				{
-					score += Math.Pow(plays[player1][player2] - averagePlays, 4);
-				}
-			}
-
-			List<List<int>> transposedGrid = TransposeGrid(grid);
-
-			//penalty for same colour
-			foreach (List<int> colour in transposedGrid)
-			{
-				int numberOfGames = colour.Count;
-				int numberOfColours = transposedGrid.Count;
-				int uniqueTeams = colour.Uniq().Count;
-				double gamesOnEachColour = gamesPerTeam / numberOfColours;
-				int penalties = colour.FindAll(team => colour.FindAll(t => t == team).Count > gamesOnEachColour).Count;
-
-				double penalty = penalties * 100000;
-				score += penalty;
-			}
-
-			foreach(List<int> row in grid)
-			{
-				int uniquePlayers = row.Uniq().Count;
-				score += Math.Abs(uniquePlayers - teamsPerGame) * 100000; // penalty for playing themselves
-			}
-
-			for (int game = 0; game < grid.Count - 1; game++)
-			{
-				foreach (var team in grid[game])
-				{
-					var nextGame = grid[game + 1];
-					if(nextGame.Contains(team))
-					{
-						var isRefTho = hasRef && (nextGame.IndexOf(team) == (nextGame.Count - 1) || grid[game].IndexOf(team) == (grid[game].Count - 1));
-						if (log) { Console.WriteLine("team " + team + " is in games " + game + " and " + (game + 1) + " back to back" + (isRefTho ? " but one is a ref game" : "")); }
-						score += BACK_TO_BACK_PENALTY - (isRefTho ? BACK_TO_BACK_PENALTY * (0.5) : 0);
-					}
-				}
-			}
-
-			return score;
-		}
-
-		List<List<int>> AddGrids(List<List<int>> grid1, List<List<int>> grid2)
-		{
-			
-			if (grid1.Count > 0 && grid1.Count == grid2.Count && grid1[0].Count == grid2[0].Count)
-			{
-				List<List<int>> grid3 = new List<List<int>>();
-				for (int i = 0; i < grid1.Count; i++)
-				{
-					List<int> row = new List<int>();
-					for (int j = 0; j < grid1[i].Count; j++)
-					{
-						if(grid2.Count > i && grid2[i].Count > j)
-						{
-							row.Add(grid1[i][j] + grid2[i][j]);
-						} else
-						{
-							row.Add(grid1[i][j]);
-						}
-					}
-					grid3.Add(row);
-				}
-				return grid3;
-			} else
-			{
-				return grid1;
-			}
-		}
-
-		List<List<int>> CalcPlays(List<List<int>> grid, bool hasRef, List<List<int>> existingPlays)
-		{
-			List<List<int>> newGrid = grid.ConvertAll(row => row.ConvertAll(cell => cell));
-			// ignore ref games
-			if (hasRef)
-			{
-				List<List<int>> transposedGrid = TransposeGrid(newGrid);
-				transposedGrid.RemoveAt(transposedGrid.Count - 1);
-				newGrid = TransposeGrid(transposedGrid);
-			}
-
-			int totalTeams = FlattenGrid(newGrid).Uniq().Count;
-			List<List<int>> plays = new List<List<int>>();
-			for(int team1 = 0; team1 < totalTeams; team1++)
-			{
-				List<int> row = new List<int>();
-				for (int team2 = 0; team2 < totalTeams; team2++)
-				{
-					if (team1 == team2)
-					{
-						List<List<int>> games = newGrid.FindAll(g => g.FindAll(t => t == team1).Count > 1);
-						row.Add(games.Count);
-					}
-					else
-					{
-						List<List<int>> games = newGrid.FindAll(g => g.Contains(team1) && g.Contains(team2));
-						row.Add(games.Count);
-					}
-				}
-				plays.Add(row);
-			}
-
-			return AddGrids(plays, existingPlays);
-		}
-
-		List<T> Shuffle<T>(List<T> list)
-		{
-			Random random = new Random();
-			int n = list.Count;
-			while (n > 1)
-			{
-				n--;
-				int k = random.Next(n + 1);
-				T value = list[k];
-				list[k] = list[n];
-				list[n] = value;
-			}
-			return list;
-		}
-
-		List<List<int>> RandomiseGrid (List<List<int>> grid)
-		{
-			for(int i = 0; i < grid.Count; i++)
-			{
-				Random random = new Random();
-				grid[i] = Shuffle(grid[i]);
-			}
-			List<List<int>> transposed = TransposeGrid(grid);
-			for (int i = 0; i < grid.Count; i++)
-			{
-				Random random = new Random();
-				grid[i] = Shuffle(grid[i]);
-			}
-			return TransposeGrid(grid);
-
-		}
-
-
-		List<List<int>> SetupGrid (FixtureTeams teams, double teamsPerGame, double gamesPerTeam)
-		{
-			int numberOfTeams = teams.Count();
-			double numGames = (numberOfTeams / teamsPerGame) * gamesPerTeam;
-
-			// set up grid
-			List<int> arr = Enumerable.Repeat(-1, (int) teamsPerGame * (int)Math.Ceiling(numGames)).ToList();
-			int index = 0;
-			List<int> updatedArr = new List<int>();
-			foreach (int el in arr)
-			{
-				updatedArr.Add((int)(index % numberOfTeams));
-				index++;
-			}
-			List<List<int>> grid = updatedArr.ChunkBy((int)teamsPerGame);
-			for (int i = 0; i < 5000; i++)
-			{
-				RandomiseGrid(grid);
-			}
-
-
-			return grid;
-
-		}
-
-		public List<List<int>> GetLeagueGrid(League league)
-		{
-			List<List<int>> grid = new List<List<int>>();
-			foreach(Game game in league.AllGames)
-			{
-				List<int> row = new List<int>();
-				foreach(GameTeam team in game.Teams)
-				{
-					row.Add(team.TeamId - 1 ?? -1);
-				}
-				grid.Add(row);
-			}
-
-			return grid;
-		}
-
-		public List<List<int>> PadPlaysToTeamNumber(int numberOfTeams, List<List<int>> plays)
-		{
-			Console.WriteLine(numberOfTeams);
-			List<List<int>> newGrid = plays.ConvertAll(row => row.ConvertAll(cell => cell));
-			int teamPlays = newGrid.Count;
-			if (numberOfTeams > teamPlays)
-			{
-				for(int i = 0; i < teamPlays; i++)
-				{
-					for(int j = teamPlays; j < numberOfTeams; j++)
-					{
-						newGrid[i].Add(0);
-					}
-				}
-				for(int i = teamPlays; i < numberOfTeams; i++)
-				{
-					List<int> arr = Enumerable.Repeat(0, numberOfTeams).ToList();
-					newGrid.Add(arr);
-				}
-
-				return newGrid;
-			} else
-			{
-				return plays;
-			}
-		}
-
-		int TeamsPerGame()
-		{
-			return (red.Checked ? 1 : 0) + (blue.Checked ? 1 : 0) + (green.Checked ? 1 : 0) + (yellow.Checked ? 1 : 0) + (referee.Checked ? 1 : 0);
-		}
-
-		string TeamColours()
-		{
-			string colours = "";
-			colours += red.Checked ? "1," : "";
-			colours += blue.Checked ? "2," : "";
-			colours += yellow.Checked ? "4," : "";
-			colours += green.Checked ? "3," : "";
-			colours += referee.Checked ? "17," : "";
-			return colours;
-		}
-
 		void ButtonGenerateClick(object sender, EventArgs e)
 		{
 			buttonGenerate.Text = "Generating...";
@@ -589,28 +123,21 @@ namespace Torn.UI
 
 		private void GenerateTeamGrid()
 		{
+			teamGrid = new TeamGrid() { BackToBackPenalty = (double)backToBackPenalty.Value };
+
+			if (red.Checked) teamGrid.Colours.Add(Colour.Red);
+			if (blue.Checked) teamGrid.Colours.Add(Colour.Blue);
+			if (green.Checked) teamGrid.Colours.Add(Colour.Green);
+			if (yellow.Checked) teamGrid.Colours.Add(Colour.Yellow);
+			if (referee.Checked) teamGrid.Colours.Add(Colour.Referee);
+
+			teamGrid.GenerateTeamGrid(Holder.League, Holder.Fixture, fixtureSelectedTeams, (int)maxTime.Value * 1000, (int)gamesPerTeamInput.Value, gameDateTime.Value, (int)minBetween.Value);
 			continueGenerating.Enabled = false;
-			Holder.Fixture.Teams.Clear();
-			Holder.Fixture.Teams.Populate(fixtureSelectedTeams);
-			Holder.Fixture.Games.Clear();
-			double numberOfTeams = Holder.Fixture.Teams.Count;
-			bool hasRef = referee.Checked;
-			double teamsPerGame = TeamsPerGame();
-			double gamesPerTeam = (double)gamesPerTeamInput.Value + (hasRef ? (double)gamesPerTeamInput.Value / (teamsPerGame - 1) : 0);
-			int maxMillis = (int)maxTime.Value * 1000;
-
-			List<List<int>> existingGrid = GetLeagueGrid(Holder.League);
-			List<List<int>> existingPlays = CalcPlays(existingGrid, false, new List<List<int>>());
-			LogGrid(existingPlays);
-
-			List<List<int>> existingPlaysPadded = PadPlaysToTeamNumber((int)numberOfTeams, existingPlays);
-			LogGrid(existingPlaysPadded);
-
-			List<List<int>> grid = GetGrid(Holder.Fixture.Teams, teamsPerGame, gamesPerTeam, hasRef, existingPlaysPadded, maxMillis);
-
-			Holder.Fixture.Games.Parse(grid, Holder.Fixture.Teams, gameDateTime.Value, TimeSpan.FromMinutes((double)minBetween.Value), TeamColours());
 
 			RefreshReports();
+
+			scoreLabel.Text = teamGrid.ScoreText;
+			scoreLabel.BackColor = teamGrid.ScoreColor;
 		}
 
 		private void RefreshReports()
@@ -641,9 +168,9 @@ namespace Torn.UI
 			buttonGenerate.Enabled = false;
 			continueGenerating.Enabled = false;
 			int maxMillis = (int)maxTime.Value * 1000;
-			List<List<int>> grid = ContinueMixing(previousGrid, previousGamesPerTeam, previousHasRef, previousExistingPlays, maxMillis, previousBestScore);
+			List<List<int>> grid = teamGrid.ContinueMixing(maxMillis);
 
-			Holder.Fixture.Games.Parse(grid, Holder.Fixture.Teams, gameDateTime.Value, TimeSpan.FromMinutes((double)minBetween.Value), TeamColours());
+			Holder.Fixture.Games.Parse(grid, Holder.Fixture.Teams, gameDateTime.Value, TimeSpan.FromMinutes((double)minBetween.Value), teamGrid.TeamColours());
 
 			RefreshReports();
